requirement.txt
absl-py==1.4.0
addict==2.4.0
albumentations==1.3.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
conda-pack==0.6.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
docker==6.1.2
efficientnet-pytorch==0.7.1
einops==0.6.1
entrypoints==0.4
et-xmlfile @ file:///home/conda/feedstock_root/build_artifacts/et_xmlfile_1674664118162/work
fastjsonschema==2.16.3
filelock==3.12.0
fonttools==4.38.0
fsspec==2023.1.0
Generalized-Wasserstein-Dice-Loss @ git+https://github.com/LucasFidon/GeneralizedWassersteinDiceLoss.git@4aa35abb077863f58c392f493886b967d6cc3eb6
google-auth==2.18.0
google-auth-oauthlib==0.4.6
grpcio==1.54.0
huggingface-hub==0.14.1
idna==3.4
imageio==2.28.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-server==1.24.0
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.5
mmcv==2.0.0
mmcv-full==1.7.0
mmengine==0.7.4
mmsegmentation==0.30.0
munch==2.5.0
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==2.6.3
notebook==6.5.4
notebook_shim==0.2.3
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1649806299270/work
oauthlib==3.2.2
opencv-python-headless==4.7.0.72
openpyxl @ file:///home/conda/feedstock_root/build_artifacts/openpyxl_1632332968451/work
packaging==23.1
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
pretrainedmodels==0.7.4
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.15.1
pynvml==11.5.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.3
QtPy==2.3.1
qudida==0.0.4
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
segmentation-models-pytorch==0.3.2
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.6.12
tinycss2==1.2.1
torch==1.12.0+cu113
torch-tb-profiler==0.4.1
torchaudio==0.12.0+cu113
torchvision==0.13.0+cu113
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.7
yapf==0.33.0
zipp==3.15.0
